Centralizamos la lógica de borrado de cuentas en una unica ruta que distingue entre admin y usuario normal, para asi evitar rutas semi idénticas. También borramos rutas del admin que estaban obsoletas y repetidas en el blueprint

+++ src/routes/deleteData.js
@@ -174,41 +174,76 @@ router.get('/delete_message/:id_mensaje', authMiddleware, async (req, res) => {
 
 
 //Ruta para borrar la cuenta y todos sus datos
-router.get('/borrar_cuenta', authMiddleware, async (req, res) => {
+router.get('/delete_account/:id_usuario', authMiddleware, async (req, res) => {
     let conn;
+    let author;
+    let admin;
     try {
         conn = await pool.getConnection();
 
-        const id_usuario = req.user.id;
+        const id_usuario = Number(req.params.id_usuario)
 
-        await conn.query('DELETE FROM usuarios WHERE id = ?', [id_usuario]);
+        const [user_exists] = await conn.query('SELECT * FROM usuarios WHERE id = ?',[id_usuario])
 
-        await conn.query(`
-            UPDATE categorias c
-            LEFT JOIN (
-                SELECT id_categoria, COUNT(*) as count
-                FROM hilos
-                GROUP BY id_categoria
-            ) t ON c.id = t.id_categoria
-            SET c.counter = IFNULL(t.count, 0)
-        `);
+        if (user_exists.length<=0) {
+            return res.json({deleted: false, message: 'El usuario que intentas borrar no existe'})
+        }
 
-        await conn.query(`
-            UPDATE hilos h
-            LEFT JOIN (
-                SELECT id_hilo, COUNT(*) as count
-                FROM mensajes
-                GROUP BY id_hilo
-            ) m ON h.id = m.id_hilo
-            SET h.mensajes = IFNULL(m.count, 0)
-        `);
+        const [is_admin] = await conn.query('SELECT * FROM usuarios WHERE id = ? and rol = "admin"',[req.user.id])
 
-        console.log('Todo borrado con éxito');
-        return res.json({ deleted: true });
+        if (is_admin.length>0) {
+            console.log('EL usuario que intenta borrar al otro, es admin');
+            admin = true
+            
+        }else{
+            console.log('El usuario que intenta borrar al otro, no es admin');
+            admin = false 
+        }
+
+        if (req.user.id === id_usuario) {
+            console.log('EL usuario es el mismo que va a ser borrado');
+            
+            author = true
+        }else{
+            console.log('EL usuario no es el mismo que va a ser borrado');
+            
+            author = false
+        }
+
+        if (author || admin) {
+            await conn.query('DELETE FROM usuarios WHERE id = ?', [id_usuario]);
+
+            await conn.query(`
+                UPDATE categorias c
+                LEFT JOIN (
+                    SELECT id_categoria, COUNT(*) as count
+                    FROM hilos
+                    GROUP BY id_categoria
+                ) t ON c.id = t.id_categoria
+                SET c.counter = IFNULL(t.count, 0)
+            `);
+
+            await conn.query(`
+                UPDATE hilos h
+                LEFT JOIN (
+                    SELECT id_hilo, COUNT(*) as count
+                    FROM mensajes
+                    GROUP BY id_hilo
+                ) m ON h.id = m.id_hilo
+                SET h.mensajes = IFNULL(m.count, 0)
+            `);
+
+            console.log('Todo borrado con éxito');
+            return res.json({ deleted: true });
+        }else{
+            console.log('El usuario no tiene permisos sobre esta cuenta');
+            return res.json({deleted: false, message: 'No tienes permisos de borrado sobre esta cuenta'})
+            
+        }
 
     } catch (error) {
         console.log('Error:', error);
-        return res.json({ deleted: false })
+        return res.json({ deleted: false, message:'Error interno al intentar borrar' })
     } finally {
         if (conn) conn.release();
     }
